remove and sanity

+++ src/postamt/scripts/admin.py
@@ -19,46 +19,63 @@ def main():
 @click.group()
 @click.option('--debug/--no-debug', default=False, envvar="POSTAMT_DEBUG")
 @click.option('--db', default="db.sqlite", help="The postamt database file",
-              type=click.Path(exists=True, file_okay=True), envvar="POSTAMT_DB")
+              type=click.Path(file_okay=True), envvar="POSTAMT_DB")
 @click.pass_context
 def cli(ctx, debug, db):
     """Manage postamt sqlite database."""
 
-    ctx.obj['DEBUG'] = debug
-    ctx.obj['DB'] = db
-
-    # relative path
     engine = create_engine('sqlite:///{db}'.format(db=db))
     model.initialize_sql(engine)
 
+    ctx.obj['DEBUG'] = debug
+    ctx.obj['DB'] = db
+    ctx.obj['ENGINE'] = engine
 
-@cli.command()
-@click.pass_context
-def reset(ctx):
-    """Reset the whole postamt database."""
 
-    if click.confirm('Really reset db?'):
-        os.unlink(ctx.obj['DB'])
+def init_db(engine):
+    model.create_model(engine)
 
-        engine = create_engine('sqlite:///{db}'.format(db=ctx.obj['DB']))
-        model.initialize_sql(engine)
+    # create local domain
+    try:
+        local_domain = model.Domain.query.filter(model.Domain.id == 0).one()
 
-        # create local domain
+    except NoResultFound:
         local_domain = model.Domain(id=0)
         model.DBSession.add(local_domain)
 
+    try:
+        nul_address = model.Address.query.filter(model.Address.id == 0).one()
+
+    except NoResultFound:
         null_address = model.Address(
             id=0, domain=local_domain, localpart="\0", active=False, rclass=None, transport=None)
         model.DBSession.add(null_address)
 
-        model.DBSession.commit()
+    model.DBSession.commit()
+
+
+@cli.command()
+@click.option("--reset", "-r", default=False, is_flag=True)
+@click.pass_context
+def init(ctx, reset):
+    """Init or reset the whole postamt database."""
+
+    if os.path.isfile(ctx.obj['DB']) and reset and click.confirm('Really reset db?'):
+        os.unlink(ctx.obj['DB'])
+
+    if not os.path.isfile(ctx.obj['DB']):
+        init_db(ctx.obj['ENGINE'])
+        click.echo("Database initialized: {}".format(ctx.obj['DB']))
 
 
 @cli.group()
 @click.pass_context
 def domain(ctx):
     """Manage Domain table."""
-    pass
+
+    # implicitelly init db if not existing
+    if not os.path.isfile(ctx.obj['DB']):
+        ctx.invoke(init)
 
 
 @domain.command(name="list")
@@ -83,11 +100,12 @@ def domain_add(ctx, domain_name, active, klass, rclass):
     """Add a new doman."""
 
     # test if already there
-    try:
-        domain = model.Domain.query.filter(model.Domain.name == domain_name).one()
+    domain = model.Domain.find(domain_name)
+
+    if domain:
         click.echo("Domain already exists: {}".format(domain_name))
 
-    except NoResultFound:
+    else:
         domain = model.Domain(name=domain_name, active=active, klass=3, rclass=20)
         model.DBSession.add(domain)
 
@@ -100,15 +118,49 @@ def domain_add(ctx, domain_name, active, klass, rclass):
 
         model.DBSession.commit()
 
-        click.echo("Added domain: {}".format(domain))
-
+        click.echo("Domain added: {}".format(domain))
 
     return domain
+
+
+@domain.command(name="rm")
+@click.argument("domain_name")
+@click.pass_context
+def domain_remove(ctx, domain_name):
+    """Remove a domain."""
+
+    domain = model.Domain.find(domain_name)
+
+    if domain is None:
+        click.echo("Domain not found: {}".format(domain_name))
+
+    else:
+        if not domain.addresses and not domain.transports:
+
+            model.DBSession.delete(domain)
+            model.DBSession.commit()
+
+            click.echo("Domain removed: {}".format(domain_name))
+
+        else:
+
+            click.echo("There are addresses or transports left - remove them first explicitelly!")
+            for address in domain.addresses:
+                click.echo("Address: {}".format(address))
+
+            for transport in domain.transports:
+                click.echo("Transport: {}".format(transport))
+
 
 @cli.group()
 @click.pass_context
 def address(ctx):
     """Manage Address table."""
+
+    # implicitelly init db if not existing
+    if not os.path.isfile(ctx.obj['DB']):
+        ctx.invoke(init)
+
 
 
 @address.command(name="list")
@@ -132,16 +184,17 @@ def address_add(ctx, address_name, active, transport, rclass):
     """Add an address."""
     localpart, domain_name = address_name.split("@", 1)
 
-    # domain lookup
-    domain = ctx.invoke(domain_add, domain_name=domain_name)
+    # domain lookup and default add
+    domain = model.Domain.find(domain_name)
+    if not domain:
+        domain = ctx.invoke(domain_add, domain_name=domain_name)
 
-    try:
-        address = model.Address.query\
-            .filter(model.Address.localpart == localpart, model.Address.domain == domain).one()
+    address = model.Address.find(address_name)
 
+    if address:
         click.echo("Address already exists: {}".format(address))
 
-    except NoResultFound:
+    else:
         address = model.Address(localpart=localpart, domain=domain, active=active, rclass=rclass)
 
         model.DBSession.add(address)
@@ -152,10 +205,46 @@ def address_add(ctx, address_name, active, transport, rclass):
     return address
 
 
+@address.command(name="rm")
+@click.argument("address_name")
+@click.pass_context
+def address_remove(ctx, address_name):
+    """Remove an address."""
+
+    localpart, domain_name = address_name.split("@", 1)
+
+    address = model.Address.find(address_name)
+
+    if address is None:
+        click.echo("Address not found: {}".format(address_name))
+
+    else:
+
+        if not address.sources and not address.targets:
+            model.DBSession.delete(address)
+            model.DBSession.commit()
+
+            click.echo("Address removed: {}".format(address_name))
+
+        else:
+            click.echo("There are still aliases around, which relate to this address -"
+                       " please remove them first!")
+
+            for as_source in address.sources:
+                click.echo("Source for: {}".format(as_source))
+
+            for as_target in address.targets:
+                click.echo("Target for: {}".format(as_target))
+
+
 @cli.group()
 @click.pass_context
 def alias(ctx):
     """Manage Alias table."""
+
+    # implicitelly init db if not existing
+    if not os.path.isfile(ctx.obj['DB']):
+        ctx.invoke(init)
 
 
 @alias.command(name="list")
@@ -177,19 +266,23 @@ def alias_list(ctx):
 def alias_add(ctx, source_address, target_address, active):
     """Add an alias."""
 
-    # domain lookup
-    source_address = ctx.invoke(address_add, address_name=source_address)
-    target_address = ctx.invoke(address_add, address_name=target_address)
+    source = model.Address.find(source_address)
+    if not source:
+        source = ctx.invoke(address_add, address_name=source_address)
+
+    target = model.Address.find(target_address)
+    if not target:
+        target = ctx.invoke(address_add, address_name=target_address)
 
     try:
         alias = model.Alias.query\
-            .filter(model.Alias.address == source_address,
-                    model.Alias.target == target_address).one()
+            .filter(model.Alias.address == source,
+                    model.Alias.target == target).one()
 
         click.echo("Alias already exists: {}".format(alias))
 
     except NoResultFound:
-        alias = model.Alias(address=source_address, target=target_address, active=active)
+        alias = model.Alias(address=source, target=target, active=active)
 
         model.DBSession.add(alias)
         model.DBSession.commit()
@@ -197,10 +290,39 @@ def alias_add(ctx, source_address, target_address, active):
         click.echo("Alias added: {}".format(alias))
 
 
+@alias.command(name="rm")
+@click.argument("source_address")
+@click.argument("target_address")
+@click.pass_context
+def alias_remove(ctx, source_address, target_address):
+    """Remove an alias."""
+
+    source = model.Address.find(source_address)
+    target = model.Address.find(target_address)
+
+    try:
+        alias = model.Alias.query\
+            .filter(model.Alias.address == source,
+                    model.Alias.target == target).one()
+
+    except NoResultFound:
+        click.echo("There is not such alias: {} -> {} ".format(source_address, target_address))
+
+    else:
+        model.DBSession.delete(alias)
+        model.DBSession.commit()
+
+        click.echo("Alias removed: {} -> {}".format(source_address, target_address))
+
+
 @cli.group()
 @click.pass_context
 def user(ctx):
     """Manage VMailbox table."""
+
+    # implicitelly init db if not existing
+    if not os.path.isfile(ctx.obj['DB']):
+        ctx.invoke(init)
 
 
 @user.command(name="list")
@@ -243,3 +365,24 @@ def user_add(ctx, address_name, active, uid, gid, home, password):
 
         click.echo("User added: {}".format(user))
 
+
+@user.command(name="rm")
+@click.argument("address_name")
+@click.pass_context
+def user_remove(ctx, address_name):
+    localpart, domain_name = address_name.split("@", 1)
+
+    try:
+        user = model.VMailbox.query\
+            .join(model.Address, model.Domain)\
+            .filter(model.Address.localpart == localpart)\
+            .filter(model.Domain.name == domain_name)\
+            .one()
+
+        model.DBSession.delete(user)
+        model.DBSession.commit()
+
+        click.echo("User removed: {}".format(address_name))
+
+    except NoResultFound:
+        click.echo("User not found: {}".format(address_name))
